Prepare integrationtest directory.
